chore(deps): bump graphql to 17 and graphql-codegen to v6/v7
Consolidates four dependabot bumps for the indexer-public-data-provider
package into one coordinated major upgrade:

- graphql 16.14.0 -> 17.x
- @graphql-codegen/cli 6 -> 7
- @graphql-codegen/client-preset 5 -> 6
- @graphql-codegen/typescript-operations 5 -> 6
- @graphql-codegen/typescript 5 -> 6 (missed by dependabot; required by
  client-preset 6)

client-preset 6 drops the `typescript` plugin from its `graphql.ts` output,
so schema output types (RegularTransaction, ContractBalance, Segment,
TransactionResult) and the Maybe/InputMaybe helpers are no longer emitted
there. Restore them via a dedicated `typescript`-plugin output
(src/gen/schema-types.ts) and repoint the affected imports.

graphql-ws 6 still declares a `^15.10.1 || ^16` peer on graphql, but the
indexer only uses the graphql-ws client (createClient), which has no runtime
graphql dependency; the unmet-peer warning has no runtime effect.

Verified: yarn gen, build, tsc (turbo check), eslint, and 216 unit tests all
pass.
