+++ packages/indexer-public-data-provider/package.json
@@ -35,17 +35,17 @@
     "@midnight-ntwrk/midnight-js-utils": "workspace:*",
     "buffer": "^6.0.3",
     "cross-fetch": "^4.1.0",
-    "graphql": "^16.14.0",
+    "graphql": "^17.0.1",
     "graphql-ws": "^6.0.8",
     "isomorphic-ws": "^5.0.0",
     "rxjs": "^7.8.2",
     "ws": "^8.20.0"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^6.0.0",
-    "@graphql-codegen/client-preset": "^5.0.0",
-    "@graphql-codegen/typescript": "^5.0.0",
-    "@graphql-codegen/typescript-operations": "^5.0.8",
+    "@graphql-codegen/cli": "^7.1.3",
+    "@graphql-codegen/client-preset": "^6.0.1",
+    "@graphql-codegen/typescript": "^6.0.2",
+    "@graphql-codegen/typescript-operations": "^6.0.5",
     "@types/ws": "^8.18.1",
     "vitest": "^4.1.7"
   },
